+++ lib/solidus_extension_dev_tools/rspec/feature_helper.rb
@@ -10,21 +10,14 @@
 require 'solidus_extension_dev_tools/rspec/rails_helper'
 
 require 'capybara-screenshot/rspec'
-require 'webdrivers'
-require 'selenium/webdriver'
+require 'capybara/apparition'
 
-Capybara.javascript_driver = (ENV['CAPYBARA_DRIVER'] || :selenium_chrome_headless).to_sym
+Capybara.javascript_driver = (ENV['CAPYBARA_DRIVER'] || :apparition).to_sym
 Capybara.default_max_wait_time = 10
-Capybara.server = :webrick
-Capybara.register_driver :selenium_chrome_headless do |app|
-  Capybara::Selenium::Driver.load_selenium
-  browser_options = ::Selenium::WebDriver::Chrome::Options.new.tap do |opts|
-    opts.args << '--headless'
-    opts.args << '--disable-gpu' if Gem.win_platform?
-    opts.args << '--disable-site-isolation-trials'
-    opts.args << '--window-size=1920,1080'
-  end
-  Capybara::Selenium::Driver.new(app, browser: :chrome, options: browser_options)
+Capybara.server = :puma, { Silent: true } # A fix for rspec/rspec-rails#1897
+
+Capybara.register_driver :apparition do |app|
+  Capybara::Apparition::Driver.new(app, window_size: [1920, 1080])
 end
 
 require 'spree/testing_support/capybara_ext'